FreadTokenizer implementation moved into a separate file
